requirements.txt
beautifulsoup4
requests
amphoradata
pandas